test(testing): физически разделены AppDependencies tests
AppDependenciesTests сокращён до composition suite, остальные suites вынесены в отдельные файлы по behavior areas
общие doubles, scenarios и async ожидания перенесены в AppDependenciesTestSupport с appdependencies-префиксами
