+++ RSSReaderTests/Infrastructure/AppDependenciesSwiftDataConfigurationTests.swift
@@ -1,0 +1,51 @@
+import CloudKit
+import Foundation
+import SwiftData
+import Testing
+@testable import RSSReader
+
+@Suite("Infrastructure / AppDependencies / SwiftData Configuration")
+@MainActor
+struct AppDependenciesSwiftDataConfigurationTests {
+    @Test
+    func appDependenciesBuildSwiftDataConfigurationsWithActiveCloudKitOnlyForSyncBackedStore() throws {
+        let configurationPlan = AppDependencies.makeSwiftDataConfigurationPlan(
+            modelPartition: AppPersistenceModelPartition.current,
+            isStoredInMemoryOnly: false
+        )
+
+        let configurationsByName = Dictionary(
+            uniqueKeysWithValues: configurationPlan.modelContainerConfigurations.map { ($0.name, $0) }
+        )
+        let syncBackedConfiguration = try #require(configurationsByName["SyncBackedStore"])
+        let localOnlyConfiguration = try #require(configurationsByName["LocalOnlyStore"])
+        let syncBackedCloudKitDatabase = String(describing: syncBackedConfiguration.cloudKitDatabase)
+        let localOnlyCloudKitDatabase = String(describing: localOnlyConfiguration.cloudKitDatabase)
+
+        #expect(configurationPlan.modelContainerConfigurations.count == 2)
+        #expect(
+            syncBackedCloudKitDatabase
+                .contains("_privateDBName: Optional(\"\(CloudKitContainerConfiguration.containerIdentifier)\")")
+        )
+        #expect(localOnlyCloudKitDatabase.contains("_none: true"))
+        #expect(syncBackedConfiguration.isStoredInMemoryOnly == false)
+        #expect(localOnlyConfiguration.isStoredInMemoryOnly == false)
+    }
+
+    @Test
+    func appDependenciesBuildSwiftDataConfigurationsWithoutCloudKitWhenSyncEnablementResolvesDisabled() throws {
+        let configurationPlan = AppDependencies.makeSwiftDataConfigurationPlan(
+            modelPartition: AppPersistenceModelPartition.current,
+            isStoredInMemoryOnly: false,
+            syncBackedCloudKitPolicy: .disabled
+        )
+
+        let configurationsByName = Dictionary(
+            uniqueKeysWithValues: configurationPlan.modelContainerConfigurations.map { ($0.name, $0) }
+        )
+        let syncBackedConfiguration = try #require(configurationsByName["SyncBackedStore"])
+        let syncBackedCloudKitDatabase = String(describing: syncBackedConfiguration.cloudKitDatabase)
+
+        #expect(syncBackedCloudKitDatabase.contains("_none: true"))
+    }
+}
